requirements.txt
Flask==2.2.5
Pillow==10.2.0
reedsolo==2.0.10
